Move bitcode to full features
