+++ crates/tests/Cargo.toml
@@ -17,7 +17,7 @@ full = ["rmp-serde", "bincode", "postcard", "musli-json", "serde_json", "bitcode
 text = ["musli-json", "serde_json"]
 std = ["alloc", "musli?/std", "musli-wire?/std", "musli-storage?/std", "musli-json?/std", "musli-zerocopy?/std", "rand/std", "serde_json?/std", "rkyv?/std", "serde?/std"]
 alloc = ["musli?/alloc", "musli-wire?/alloc", "musli-storage?/alloc", "musli-json?/alloc", "musli-zerocopy?/alloc"]
-simdutf8 = ["musli-wire?/simdutf8", "musli-storage?/simdutf8", "musli-descriptive?/simdutf8", "musli-json?/simdutf8"]
+simdutf8 = ["musli-wire?/simdutf8", "musli-storage?/simdutf8", "musli-descriptive?/simdutf8", "musli-json?/simdutf8", "bitcode?/simdutf8"]
 parse-full = ["musli-json?/parse-full"]
 musli-wire = ["musli", "dep:musli-wire", "simdutf8"]
 musli-descriptive = ["musli", "dep:musli-descriptive", "simdutf8"]
@@ -32,7 +32,7 @@ test = ["musli", "musli-wire", "musli-storage", "musli-descriptive", "musli-valu
 model-no-usize = []
 model-no-cstring = []
 model-no-map = ["model-no-map-string-key"]
-model-no-map-string-key = []
+model-no-map-string-key = ["model-no-map"]
 model-no-vec = []
 model-no-tuple = []
 model-no-128 = []
@@ -45,8 +45,8 @@ postcard = ["dep:postcard", "serde"]
 rkyv = ["dep:rkyv", "std", "model-no-usize", "model-no-map-string-key", "model-no-tuple", "model-no-cstring"]
 serde_cbor = ["dep:serde_cbor", "serde", "model-no-128", "model-no-map"]
 dlhn = ["dep:dlhn", "serde", "model-no-128", "model-no-map"]
-bitcode = ["dep:bitcode", "model-no-128", "model-no-map", "serde"]
-bitcode-derive = ["bitcode", "bitcode/derive"]
+bitcode = ["dep:bitcode", "serde", "simdutf8"]
+bitcode-derive = ["dep:bitcode", "bitcode/derive", "simdutf8"]
 
 [dependencies]
 musli = { path = "../musli", version = "0.0.86", default-features = false, optional = true, features = ["alloc"] }
